% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/central.R
\name{CLT.plot}
\alias{CLT.plot}
\title{중심극한정리
중심극한정리를 카이제곱을 이용하여 직접 확인할수있다
CLT.plot(rchisq, n = 1, df = 1)
CLT.plot(rchisq, n = 30, df = 1)일때를 비교}
\usage{
CLT.plot(r.dist, n, ...)
}
\description{
중심극한정리
중심극한정리를 카이제곱을 이용하여 직접 확인할수있다
CLT.plot(rchisq, n = 1, df = 1)
CLT.plot(rchisq, n = 30, df = 1)일때를 비교
}
